Split out steps from the view
